+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+-   Added IronPython SMB Agents/Listener (@Cx01N)
 -   Expand file options to plugins, stagers, and listeners (@Vinnybod)
 
 ## [5.4.2] - 2023-06-07